bgp,rr: Implement peeringAddressFamily specification
Implement a peeringAddressFamily specification for the operator. The
operator now renders NodeConfig with one (when the peeringAddressFamily
is "IPv4Only" or "IPv6Only") or two (when the peeringAddressFamily is
"Dual") peers.

The peering address selection is based on the peeringAddressFamily and
we no longer have a "prefer IPv6" rule. The peering address family is
always explicitly specified by users or derived from the Cilium's
configuration (ipv4.enabled and ipv6.enabled) if peeringAddressFamily is
omitted.

For IPv4 peer, peerConfigRefV4 is used as a rendered peerConfigRef. For
IPv6 peer, peerConfigV6 is used as rendered peerConfigRef.

We don't have to change the agent side because regardless of the
peeringAddressFamily, we just render the regular peer and agent doesn't
have to be aware of address family.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>