privnet: correctly process all relevant network event changes
Currently, the mapentries reconciliation logic is supposed to skip
network updates if that network had already been processed once. Yet,
the filtering logic (i.e., [MapEntries.isKnownNetwork]) doesn't
actually work as intended, because it tries to match on a part of
the index key only, but [Get] expects a full match. Hence, it never
finds an entry, which means that no event is actually skipped.

Incidentally, the broken behavior is actually the correct one,
because we need to process all endpoints and routes again if
something changed in the network spec (e.g., INB, interface),
to adapt the corresponding entries accordingly. The existing
logic also takes already care of not skipping possible endpoint
and route deletion events in that case, so that possible stale
entries are correctly removed even following a network update.

Hence, let's just get rid of the extra filtering, and process
all network event changes. Still, let's try to be a bit smart,
and skip triggering the relatively expensive logic if the
network change does not actually affect the reconciler (e.g.,
in case of subnet/route changes, as these are propagated via
the dedicated table).

Fixes: 533ad05ab754 ("privnet: Add map entries reconciler")
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>