egressgwha: Fix localOnlyFirst fallback across groups
[ v1.16-ce commit 1dfa5ab6a0ce6f46bde99aa28510ffc4464da5e7 ]

Previously, when azAffinity was set to localOnlyFirst,
fallback to non-local gateways was allowed within a group
if the group itself had no local gateways in the zone,
even if other groups had local gateways in that zone.

This violated the expected behavior of localOnlyFirst,
which disallows fallback to non-local gateways in zones
where any group has local gateway nodes.

This commit ensures that fallback to non-local gateways
is only allowed when no group has local gateways in the zone.

Signed-off-by: Yusuke Suzuki <yusuke.suzuki@isovalent.com>