k8s,bgp: Make VRFRef in IsovalentBGPNodeVRF/BGPVRF  optional
If privateNetworkRef is specified, vrfRef is ignored, so make vrfRef optional.
Either vrfRef or privateNetworkRef should be provided, never both at the same time.

Since vrfRef field is now optional, it should be a pointer type
in the Go definition, and the related code should count with it being optional.

vrfRef field also can not be used in list-map-keys as it is now optional.

Signed-off-by: Rastislav Szabo <rastislav.szabo@isovalent.com>