privnet: Introduce operator status reconciler and VNI assignment
Introduce a new operator logic to reconcile ClusterwidePrivateNetwork's
status field. The only use case we have now is VNI assignment.

When the VNI is requested via spec.vni field, the operator checks
whether the requested VNI is taken by other private network. If there's
no conflict, it assigns VNI to the private network via .status.vni
field.

If there's a conflict, no private network gets the VNI and the condition
type of VNIConflict is set. If there's an existing allocation, it will
be cleared. This causes the traffic disruption. We can improve this by
introducing validation webhook for example in the future.

This commit introduces the following StateDB table changes to the operator:

- private-networks: Reflected from the k8s API, extended to reflect VNI
  allocation related fields.
- private-networks-status: New table, the desired status that is eventually
  pushed to the ClusterwidePrivateNetwork's status field.

The dataflow of VNI allocation is:

- k8s spec => private-networks
- private-networks => private-networks-status
- private-networks-status => k8s status

Co-authored-by: Rastislav Szabo <rastislav.szabo@isovalent.com>
Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>