privnet: extract and generalize the observers logic
The EndpointsObserver is currently tailored to observe private network
endpoint events from clustermesh. With the plan to introducing more
observers, let's extract the logic into a separate package, and
generalize it so that the queuing/observe bits can be easily reused.

No functional changes are performed, besides the addition of an extra
check to avoid buffering events once the observer context has been
canceled, given that they would never be observed anyways.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>