privnet: introduce nodes observer
Introduce an observer to intercept and observe node events (both from
the local and remote clusters), for use by the to-be-introduced INBs
reconciler. Specifically, the observer wraps the NodeManager interface,
performs the appropriate operations, and then calls into the corresponding
downstream method; it is intended to be injected via [cell.DecorateAll].

Furthermore, the observer caches the information about all nodes (using
a slim version, for efficiency), and allows retrieving them via by
cluster name via the [List] method. This will be useful during the
INBs reconciliation process.

Overall, this is a temporary workaround until a dedicated nodes statedb
table is introduced, as at that point we could directly observe it
rather than needing this hack. Another possibility could be having
NodeManager itself implement the [stream.Observable] interface, but
that requires additional changes on the OSS side, while the current
approach is more self-contained.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>