privnet: initial scaffolding to support automatic INBs configuration
This commit extends the privnet table with INB selectors (i.e.,
constituted by target cluster and optional node selector), to be used
for finding and automatically populating the list of INBs serving a
given network.

The CRD structure is not changed at this stage yet, and consequently
the new selectors field is not actually populated, to avoid breaking
the existing script tests until the new logic is in place.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>