ilb: never plumb LBVIP pseudo K8s service into datapath
Currently, the "pseudo" K8s service that is created per `LBVIP`
is plumbed into datapath on all nodes. This doesn't make any sense
because the only purpose of this Service is to reserve the (potentially
shared) VIP from Cilium LB IPAM and use the sharing functionality to
use the VIP for all `LBService`'s that reference the same `LBVIP`.

Therefore, this commit adds the annotation `service.cilium.io/node-selector`
to the K8s Service and configures it to never get applied on any node (neither
T1 nor T2) (No ClusterIP and Loadbalancer service gets created).

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>