privnet: initial scaffolding of the INB health checking logic
Introduce the health checker interface, which allows to register and
unregister (INB, private network) pairs for health checking, and
exposes back an observable interface with the stream of health update
events. For the moment, let's also introduce a _dummy_ implementation
that simply claims all pairs as healthy, which can be useful for
testing purposes, and matches the existing behavior (i.e., no checks).

Additionally, let's extend the INBs reconciler to register/unregister
the different pairs as well, as well as observe the stream of events
and update the state of each INB pair as appropriate.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>