privnet: reconciliation logic to promote active INB for each network
Extend the INBs reconciler with a new logic that is responsible for
promoting one of the healthy INBs for each network to active, and
replacing it in case it turns unhealthy. Currently, the logic simply
randomly picks one of the healthy INBs, but it can be extended in the
future to add some smartness towards better sharding.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>