bgp: Introduce importEVPNRouteReconciler
Introduce a new state reconciler for EVPN route import. It gets the
desired routes from BGP RIB and reconcile it against the Cilium's RIB.
The only route we support now is a "Pure" RT-5 route which is a special
case of the EVPN RT-5 which has a Router's MAC extended community which
eliminates the need for having separate RT-2 to resolve router's MAC
address (VXLAN's inner MAC address). We also have following limitations.

- ESI field must be zero
- ETag field must be zero
- GatewayIP must be zero
- Must have at least one route target
- Must have encap extended-community with VXLAN tunnel type

When we encounter the unsupported route, we simply ignore them. In the
future, we might want to expose a metrics that counts the number of
ignored routes by reasons.

Like we do for SRv6 L3VPN routes, the reconciler performs full
reconciliation for every route updates. This obviously doesn't scale. We
may want to get rid of this in the future.

The unit test covers the parsing of MP_REACH_NLRI and
EXTENDED_COMMUNITY attributes. We also have a replay test using BGPlay
with the Pure RT5 advertisement taken from Nexus9kv.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>