privnet: propagate node name and flags as part of endpoint information
In preparation for supporting high availability of INBs, let's extend
the PrivateNetworkEndpointSlice type, and the corresponding kvstore
representation, to additionally covey the name of the node hosting
the given (slice of) endpoints. This will be used by the mapentries
reconciler to filter out the entries advertised by the INBs nodes
currently not selected as active. While we could currently infer
it from the name of the ipnes resource, that's just an implementation
detail that may change in the future, so let's be explicit with
a dedicated field.

Similarly, let's also associate to each endpoint a set of flags,
currently only conveying whether a given endpoint is external. We
consider an endpoint to be external in case access to it is provided
via an INB. Again, this information is key to performing filtering
for supporting HA of INBs, without having to make assumptions that
INB nodes cannot host normal workloads, that may prove to be limiting
in the future.

Update the appropriate reconcilers to correctly set and propagate
the newly introduced bits of information.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>