loader: retry with xdp frags
[ upstream commit 62f4856809eeac764776de1588d7fba43affbb35 ]

The commit adds retry logic in case the XDP program loading failed with
the `invalid argument` error. The error might indicate that the network
interface is configured with a jumbo MTU, so we can retry loading after
setting the `BPF_F_XDP_HAS_FRAGS` flag and hope that the NIC driver is XDP
Fragment aware.

Signed-off-by: viktor-kurchenko <viktor.kurchenko@isovalent.com>
Signed-off-by: Timo Beckers <timo@isovalent.com>