cni: add option to setup Pod netns with tcp mtu probing.
[ upstream commit 9fdca1d3dd52df695d89c887870f3fa0752ec7ac ]

[ Notes: applying again the configmap hunk, as it looks like
  that these changes got incorrectly reverted by c47c16c639e6
  ("Merge remote-tracking branch 'upstream/main' into
  pr/pippolo/main-sync-2025-11-18") ]

This adds option for enabling Linux TCP PMPMTU probing by default for
Cilium managed endpoints via enabling the tcp_mtu_probing
sysctl option on Pod netns.
By default, all endpoints will now have this option enabled.
When on, this will configure the endpoint with reasonable config
values for base mss, as well as setting the mtu probing value
to be '2'; which means it will always attempt using mtu probing.

The idea of this change is that we will allow for more robust
MTU handling at a very small cost of sub-optional initial MTU
windows.

In the future, we may want to also be more smart with regards to
how we set the base MSS in combination with vxlan/geneve tunneling.

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>