privnet: rework private networks caching in mapentries reconciler
Starting from [1], the mapentries reconciler already caches the known
private networks to avoid unnecessarily repeating work in case none
of the relevant parameters changed. In preparation for the subsequent
reconciliation of active INB information as well, let's formalize this
a bit more, defining a dedicated SlimPrivateNetwork type, intended to
contain only the necessary information (including the details of the
active INB, which will come from a different table). Additionally,
let's make use of this cache (rather than directly querying the table)
from the other components of the mapentries reconciler as well, for
consistency, and because it is more efficient.

[1]: b54f2c76fd25 ("privnet: correctly process all relevant network event changes")

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>