privnet: extend mapentries reconciler to handle INB changes
Extend the mapentries reconciler to additionally react upon active INB
changes, so that the associated entries can be updated accordingly.

Rather than leveraging a changes iterator, which would wake for any
(possibly unrelated) change in the INBs table, the reconciler explicitly
watches the individual watch channels associated with the active INB
entry of any known private network. Additionally, it makes use of a
map to link each channel to the corresponding network(s), in order
to be able to synthetize the necessary events without having to loop
over all networks every time.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>