ilb: validate and auto-evaluate in case of ip family mismatch
While T1&T2 (with Envoy) supports different ip address families on the
frontend (LBVIP&LBService) and the LBBackendPool, T1-only mode doesn't
support this. The reason is that Envoy always acts as a proxy that
establishes a new connection to the upstream backend.

Therefore, this commit validates and auto-evaluates the deployment mode
based on the ip families configured on the frontend (LBVIP & LBService)
and the configured IP addresses on the referenced LBBackendPools.

Only if every referenced LBBackendPool of Type "IP" defines at least
one address per address family that is exposed on the LBService - the
deployment mode can be of type `t1-only`. Otherwise it defaults to `t1-t2`.

If `t1-only` is enforced via `forceDeploymentMode` (for TCP&UDP proxy), the
validation fails if above condition isn't fulfilled.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>