ilb: fix deployment mode evaluation in cilium lb status
Currently, the status columns that display information about
T2, HC T2->Backends and backends conditionally changes dependent
on the deployment mode of the LBService (t1-only or t1-t2).

This logic still builds on top of the field `.forceDeploymentMode` (on
TCP & UDP Proxy) even though the field with the "evaluated" mode has been added to the
status in the meantime (`.status.applications.[tcpProxy|udpProxy].deploymentMode`).

This commit refactors the logic and only uses the spec field as fallback if the status
field isn't yet populated by the ILB controlplane.

Note: The column that displays the `Type` already uses the status field.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>