bpf: egressgwha: don't attempt to redirect local traffic in SEGW mode
Locally-originating traffic on the SEGW node should not be managed by the
EGW policy.

Currently we rely on identity and skb->mark checks in the to-netdev path
to exclude local traffic from being matched against the SEGW catch-all
policy. But we can go one step further - and just not include any of the
code that matches packets against EGW policy, stashes the gateway selection
into the CT map etc.

This is a lot more robust (we don't rely on netfilter to mark local
traffic), and optimizes out a bunch of unnecessary BPF code.

Signed-off-by: Julian Wiedmann <jwi@isovalent.com>