evpn: Introduce vni-mappigs reconciler
Introduce a new private-networks => vni-mappings reconciler. We cannot
rely on the Reconciler infrastructure for this case because we have a
case that the deallocation and update of the VNI. In such a case, we
need to find the old mapping by NetworkID and remove that, but from the
Update/Delete callback, we don't have a straightforward way to get them.

Co-authored-by: Rastislav Szabo <rastislav.szabo@isovalent.com>
Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>