ilb: add ipv6 e2e tests
This commit adds a set of new ILB E2E tests that test the IPv6
functionality.

The tests check TCP- & UDPProxy functionality in all combinations
of VIP address family and backend address family.

* TestTCPProxyIPv6VIPIPv6BackendT1Only
* TestTCPProxyIPv6VIPIPv6BackendT1T2
* TestTCPProxyIPv6VIPIPv6BackendAuto
* TestTCPProxyIPv6VIPIPv4BackendT1T2
* TestTCPProxyIPv6VIPIPv4BackendAuto
* TestTCPProxyIPv4VIPIPv6BackendT1T2
* TestTCPProxyIPv4VIPIPv6BackendAuto

* TestUDPProxyIPv6VIPIPv6BackendT1Only
* TestUDPProxyIPv6VIPIPv6BackendT1T2
* TestUDPProxyIPv6VIPIPv6BackendAuto
* TestUDPProxyIPv6VIPIPv4BackendT1T2
* TestUDPProxyIPv6VIPIPv4BackendAuto
* TestUDPProxyIPv4VIPIPv6BackendT1T2
* TestUDPProxyIPv4VIPIPv6BackendAuto

Note: All existing tests still use the IPv4
address to call the VIP & configure the backend pools.
In the future we might want to change that so that
all tests can be executed as IPv4 or IPv6.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>