privnet: fix release of private network IDs
Currently, the release of no longer used private network IDs is
performed as part of the [TransformMany] reconciler function.
However, this function is not invoked on old entries upon a full
relist, hence possibly leaking IDs. While this is expected to
be never a problem in practice, given the huge ID space, let's
get it fixed, moving the release task to a dedicated job. While
being there, let's also change the reconciler back to [Transform].

Fixes: 16937f40cc63 ("privnet: introduce the private networks reconcilers")
Reported-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>