ilb: add a Health Status panel to the dashboards
Adds a "Heath Status" panel to the Isovalent Load Balancer dashboard.

Relies on the set of Envoy "cluster membership" metrics, so included for
the Tier 2 HTTP and Tier 2 TCP sections of the dashboard.

Displays Total, Healthy, Unhealthy, and Offline time series plots.

The panel has the following description text:

The total number of healthy and unhealthy health checks for the service
as observed at the Tier 2 layer. The total number will most often be the
number of Tier 2 nodes multiplied by the number of Backend Pools. For
example, a service with 10 backend pools running on a cluster with 3
Tier 2 nodes will have a total of 30 health checks that should be
healthy at any given time.

Signed-off-by: Andrew Sauber <andrew.sauber@isovalent.com>