rib: Move restoreRoutes to Start hook
Reading from BPF map on Invoke hook is not safe as the map may not be
initialized at that point. Use Start hook instead. The ordering of the
restoration and the owners start to write to it depends on the
dependency order. Since route owner must depend on the RIB, it is safe
to assume the restoration will be finished before any owner start
writing.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>