rib: Fix best path calculation bug
When we compare the best path, the route with longer AD should always
lose, but in the current implementation, when the route has longer AD
than the current best route, but has smaller Owner, it wins the best
path. Fix it.

We meant to catch this case with `TestRIB_selectBestPath`, but the test
case was bad (the route with longer AD also had bigger Owner as well).
Change test case to catch this case.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>