ent/olm: drop toolchain directive
Match the root go.mod file, which also does not specify the toolchain
directive, to avoid constraining it to a specific patch version.
Specifically, this can be problematic in case it gets set to a different
value from the one used in GitHub actions, resulting in errors. For
instance, this happened in the "Generate k8s API" test, which failed with:

  go: download go1.25.2: golang.org/toolchain@v0.0.1-go1.25.2.linux-amd64: verifying module: golang.org/toolchain@v0.0.1-go1.25.2.linux-amd64: open /go/pkg/sumdb/sum.golang.org/latest: permission denied

Still, let's bump the go version to v1.25 for consistency with the
previously toolchain value.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>