privnet: add gratuitous ARP/ND cell
The cell provides a reconciler which watches map entries and sends
gratuitous ARP/ND packets for endpoint entries on the INB towards the
target network. It is only active on the INB cluster acting as a bridge.

Compared to prior versions of the code, the reconciler is only started
on the INB. Previously, it always ran the reconciler but skipped any
updates if not in INB mode.

Co-authored-by: David Bimmler <david.bimmler@isovalent.com>
Co-authored-by: Fabian Fischer <fabian.fischer@isovalent.com>
Co-authored-by: Harsimran Pabla <hpabla@isovalent.com>
Co-authored-by: Marco Iorio <marco.iorio@isovalent.com>
Co-authored-by: Sebastian Wicki <sebastian@isovalent.com>
Signed-off-by: Tobias Klauser <tobias@isovalent.com>