hubble-timescape: Remove trailing dot in push URIs
Remove the trailing dots in URIs used to push events to Timescape
such as in Vector and the Hubble Timescape exporter.

Vector does not strip the trailing dot when forwarding to openssl
and since we removed the trailing dots in generated certificates
in a past commit, a "Host mismatch" error is thrown.

The Hubble Timescape exporter part of CIlium handles that
correctly though, but let's be consistent for all URIs.

Signed-off-by: Alexandre Barone <abalexandrebarone@gmail.com>