hubble: Remove deprecated and legacy extraConfig flags
In v1.18, we deprecated using extraConfig to provide flags for the
enterprise Hubble exporter in-favor of using proper Helm options
documented in the values file under `hubble.export.static`.

In v1.19, we remove the capability to set these flags but we keep
a minimal set of configs in-order to be able to set default values
when Integrated Timescape is enabled. We also remove the checks
around already set flags in extraConfig and add helm validation
that will fail if any flag is set.

Signed-off-by: Alexandre Barone <abalexandrebarone@gmail.com>