Remove CiliumMesh
The main motivation for the removal that it's going to be replaced by
INB + PrivNet, and that we still keep spending resources in reviewing
PRs relevant to CiliumMesh and investigating CiliumMesh CI flakes.

After ther removal, some datapath hooks are no longer used. We will
consider removing them in OSS separately from this PR.

I did not sync enterprise/olm/manifests, as there is a scheduled job to
do so, as noted in enterprise/olm/README.md:

    - A scheduled workflow, [OLM
      Sync](../../.github/workflows/enterprise-olm-sync.yaml) runs
      weekly to synchronize the manifest files used by CLife with
      changes made to the helm manifests. This workflow creates a pull
      request with the required changes.

Signed-off-by: Martynas Pumputis <martynas@isovalent.com>