privnet/health: introduce the gRPC API for node to INB communication
Introduce the definition of the synchronous gRPC API used by workload
nodes to interact with INBs. More specifically, it includes a Health
service, which allows to probe the INB for healthiness, and a Networks
service, that can be used to watch the private networks served by each
INB, as well as requesting an INB to step-up as active or step-down
for a given network.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>