privnet/health: implement the INB network-related logic
This commit implements a component responsible for retrieving the list
of networks served by a given INB, and possible subsequent changes, and
bubble them up via a dedicated channel. Additionally, it tracks registered
networks and handles activation and de-activation request.

At a high level, it leverages the newly introduced gRPC Networks service,
watching for network status changes via the watch stream. Activation and
deactivation requests are enqueued and processed via a dedicated loop;
additionally, activations and deactivations are also automatically
replayed if the state reported by the INB does not match the one expected
locally, in order to fix possible divergences following e.g., a restart.

Overall, this component is intended to be started after that an INB is
detected as healthy, and stopped when it turn unhealthy. At that point,
the dedicated [ResetToUnknown] function shall be invoked to reset the
state of all networks to unknown, and emit the corresponding events.
Eventually, it can be restarted when the INB turns healthy again.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>