privnet/health: implement the gRPC-based health checker
Building on top of the prober and networking components, let's implement
the [health.Checker] interface based on the newly introduced gRPC API.

At a high level, the implementation tracks the set of target INBs,
probing for each the health status and the list of served networks, and
relaying the health events as appropriate. The health checker is
considered to be synchronized once it either received the initial
snapshot of served networks  for every INB initially registered, or
the timeout expired, as a circuit breaker in case at least one of
the INBs is unhealthy.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>