privnet/health/test: introduce serverpool implementation
In order to facilitate the testing of the health checker implementation,
let's introduce the ServerPool cell, which allows to create and control
a set of servers via script commands. In detail, the commands allow to:

* Create a new health server;
* Start and stop existing servers;
* Simulate a blocked server, which hangs on all requests;
* Configure and update the set of networks served by the INB;
* Validate the number of currently open streams.

The list of activated networks for each INB can be easily inspected via
the dedicated test-active-networks statedb table.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>