privnet/health/test: introduce checkerpool implementation
In order to facilitate the testing of the health checker implementation,
as well as that of the server-side implementation to be introduced in
future commits, let's introduce the CheckerPool cell, which allows to
create and control a set of checkers (leveraging the real implementation)
via script commands. More in detail, dedicated script commands allow to
create a new health checker and invoke the methods of the health.Checker
interface. The observed events are propagated into a dedicated, per
instance, statedb table, ready to be inspected with the usual mechanisms.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>