helm: Fix missing client certs when Timescape StreamAPI is enabled
When integrated timescape and StreamAPI are enabled, we hardcode
mTLS as enabled in the Timescape configuration by specifying the TLS
CA certificate. We also need to configure the client certificates on the
Cilium side when we configure the Hubble timescape exporter otherwise
it will complain that certificates are required upon connecting.

This was missed because mTLS support for the StreamAPI was added in
Timescape 1.8.2 but this change was introduced first with 1.8.1.

Signed-off-by: Alexandre Barone <abalexandrebarone@gmail.com>