privnet: introduce the privnet-inb-active-networks statedb table
Introduce the definition of the privnet-inb-active-networks statedb
table, which is designed to store the information about the networks
activated by each given workload node. It takes effect on INB nodes
only, and it is intended to be populated by the health checker
server-side implementation, according to node healthiness and
activation/deactivation requests.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>