privnet/health: server implementation of the health service
Scaffold the initial version of the health gRPC server, implementing the
Health service only for the moment. More specifically, the [Probe] method
implementation is responsible for replying to the workload node probes,
and marking the node as unhealthy if the subsequent probe is not received
within the expected timeout. In that case, all previously activated
networks for that node are automatically withdrawn.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>