privnet/health: server implementation of the networks service
Implement the server-side of the Networks gRPC service. Specifically,
the [Watch] method watches for changes in the private networks and
INB active networks statedb tables, and emits events accordingly.
The [Activate] and [Deactivate] methods, instead, update the active
networks table as appropriate, after validating that all preconditions
are correctly satisfied. Specifically, a network can be activated only
if served, and the requesting node is deemed healthy.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>