privnet/health/test: introduce the connection factory
In order to facilitate the testing of the health checker implementation,
let's introduce the ConnFactory cell, which coordinates the creation of
both client connections and server listeners. For convenience, servers
listen to unix sockets (with predefined names), rather than TCP ports,
and the client connection factory is configured to appropriately connect
to them.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>