privnet/health/test: enable simulating checker failures
Extend the connection factory and checkerpool testing facilities to
allow simulating checker failures --- i.e., blocking the streaming
APIs --- via a dedicated script command. More in detail, this is
implemented by means of a stream interceptor; streams are mapped
to the corresponding checker instance via the [Self] property of
the request messages. The interceptors are only injected when
explicitly enabled, to prevent introducing unintentional behavioral
differences in the other tests.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>