privnet: extend mapentries reconciler to handle ActiveNetworks changes
On INBs, we want to perform L2 announcements on the egress facing
interface (i.e., replying to ARP/ND requests, as well as sending
gratuitous ARP and ND packets) only for endpoints hosted by nodes
that promoted the current INB as active.

In order to achieve this, let's add a dedicated field to [MapEntries],
and extend the corresponding reconciler to additionally react upon
privnet-inb-active-networks changes, retriggering the processing of
all matching endpoints. Upon processing, the new [L2Announce] field
gets populated based on whether the hosting node promoted us as
active for the given network. Finally, let's introduce a new script
test to validate the correctness of the behavior.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>