privnet: adapt gneigh reconciler to make use [MapEntryRouting.L2Announce]
Update the gneigh reconciler to rely on [MapEntryRouting.L2Announce]
to determine whether a gratuitous ARP/ND should be sent for the given
endpoint, rather than inferring it based on the cluster name. The
script test is updated accordingly, leveraging the checkerpool facility
to register and activate the appropriate networks, and validate that
a gratuitous ARP/ND packets are sent only for active endpoints.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>