privnet: register the gRCP based health cell instead of the dummy one.
Replace the dummy health checker cell with the real one, and adapt the
tests accordingly. More specifically, we need to register the testing
infrastructure (i.e., connection factory, checker pool and server pool),
as well as extend the test files, where appropriate, to configure and
start the fake health servers. The INBs test is also slightly modified
to prevent a possible source of flakiness that was not present before.
The [Dummy] health checker implementation is removed, together with the
associated cell, as no longer used.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>