gha: fix build-go-caches workflow
This workflow always failed since it got enabled on main-ce in [1], with
an error while attempting to build the cilium-dnsproxy image:

> docker buildx create --platform linux/amd64,linux/arm64 --buildkitd-flags '--debug' > .buildx_builder
> /bin/sh: 1: docker: not found

That hints to the fact that docker is being invoked inside the builder
image, which is not expected. Let's get it fixed by correctly invoking
the Makefile target building the binary, instead of the one building
the docker image. This matches the behavior for the other components.

[1]: 6e772ec79f9c ("(cee) gh: fix up s/main/main-ce")

Fixes: 5ce722d532f9 ("Merge remote-tracking branch 'upstream/main' into pr/jibi/main-sync-24-09-30")
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>