bpf: egressgwha: forward & SNAT even when all GW nodes are quarantined
When an EGW node processes requests by a non-local client, it applies EGW
policy twice. First in `from-overlay` to select the egress interface, and
then in `to-netdev` to select the egress IP. In both cases the EGW policy
lookup happens within egress_gw_ha_snat_needed(). Here we currently
sanity-check whether the policy selects any GW nodes - and if not, drop
all traffic.

This breaks connectivity for established connections, when all selected
GW nodes are unavailable. In particular it affects the scenario where the
GW nodes are quarantined, and meant to support all remainining connections
until they have drained out.

But this check is actually not useful - the GW node doesn't need any
information from the node-list, and can simply SNAT all traffic that
comes in and matches the policy.

Note that egress_gw_ha_policy_entry_is_excluded_cidr() correctly
double-checks that the gateway-count is 1, before looking at the first
entry of the list.

This issue doesn't affect packets that leave a worker node (when to-netdev
applies EGW policy via egress_gw_handle_request()), since the gateway for
an established connection is obtained from the EGW-CT map (and not the
EGW policy map entry).

The same problem for the reply path has been addressed in
https://github.com/isovalent/cilium/pull/8475.

This issue doesn't affect SEGW mode, since there the default policy entry
simply hard-codes the GW node and is never updated.

Fixes: #8427
Reported-by: Yusuke Suzuki <yusuke.suzuki@isovalent.com>
Signed-off-by: Julian Wiedmann <jwi@isovalent.com>