privnet: validate that the INB egress interface is operational up
With reference to the egress interface configured on the INBs via
the ClusterwidePrivateNetwork resource, we currently validate that
it exists, and it is selected by Cilium, according to the devices
controller logic. Let's extend this to also explicitly enforce that
its operational status is up before letting the given INB serve
that network. In turn, this allows workload nodes to avoid selecting
an INB that does not have connectivity to the external network, and
trigger a failover to a different one in case of changes.

For completeness, it is worth mentioning that the devices controller
currently already partially checks that the interface is up (i.e.,
via the IFF_UP flag only) to determine if an interface is selected.
However, that check, along others, is skipped in case the interface
is explicitly matched by the user filter, hence not having the
intended effect in the private networks context.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>