privnet: visually show interface errors in private-networks table
Add an exclamation mark close to the interface name in the default
private-networks table visualization, so that it is possible to notice
at a glance if the configured interface is not able to serve traffic.
Detailed information about the error can be retrieved via the json
or yaml representations of the corresponding entry, as before.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>