cilium-cli/enterprise: rename and move ParsePolicyYAML
ParsePolicyYAML is not only used to parse policies but k8s objects from
YAML in general. Rename the function accordingly and move it to the more
appropriate utils package such that it can be easily reused without
depending on the connectivity/check package. Also inline the scheme
parameter which is always using the same value.

Signed-off-by: Tobias Klauser <tobias@isovalent.com>